Update README with new shortcuts and UI
